+++ TShockAPI/GetDataHandlers.cs
@@ -2563,7 +2563,7 @@ namespace TShockAPI
 
 			args.Player.ActiveChest = id;
 
-			if (TShock.CheckTilePermission(args.Player, x, y) && TShock.Config.RegionProtectChests)
+			if (!args.Player.HasBuildPermission(x, y) && TShock.Config.RegionProtectChests)
 			{
 				args.Player.SendData(PacketTypes.ChestOpen, "", -1);
 				return true;
@@ -2603,13 +2603,13 @@ namespace TShockAPI
 			if (OnSignEvent(args.Player, args.Data, id, x, y))
 				return true;
 
-			if (TShock.CheckTilePermission(args.Player, x, y))
+			if (!args.Player.HasBuildPermission(x, y))
 			{
 				args.Player.SendData(PacketTypes.SignNew, "", id);
 				return true;
 			}
 
-			if (TShock.CheckRangePermission(args.Player, x, y))
+			if (!args.Player.IsInRange(x, y))
 			{
 				args.Player.SendData(PacketTypes.SignNew, "", id);
 				return true;
@@ -2928,8 +2928,8 @@ namespace TShockAPI
 			}
 
 			if (args.Player.IsBouncerThrottled() ||
-				TShock.CheckTilePermission(args.Player, x, y, true) ||
-				TShock.CheckRangePermission(args.Player, x, y))
+				!args.Player.HasPaintPermission(x, y) ||
+				!args.Player.IsInRange(x, y))
 			{
 				args.Player.SendData(PacketTypes.PaintTile, "", x, y, Main.tile[x, y].color());
 				return true;
@@ -2972,8 +2972,8 @@ namespace TShockAPI
 			}
 
 			if (args.Player.IsBouncerThrottled() ||
-				TShock.CheckTilePermission(args.Player, x, y, true) ||
-				TShock.CheckRangePermission(args.Player, x, y))
+				!args.Player.HasPaintPermission(x, y) ||
+				!args.Player.IsInRange(x, y))
 			{
 				args.Player.SendData(PacketTypes.PaintWall, "", x, y, Main.tile[x, y].wallColor());
 				return true;
@@ -3322,7 +3322,7 @@ namespace TShockAPI
 				return true;
 			}
 
-			if (TShock.CheckRangePermission(args.Player, (int)position.X, (int)position.Y))
+			if (!args.Player.IsInRange((int)position.X, (int)position.Y))
 			{
 				return true;
 			}